tree: nuke TODO file
We don't do a good job of keeping this file up to date. It might end up
doing more harm than good if it confuses newcomers.

Anyway, many of these are either no longer relevant or already
implemented. Let's agree that the canonical source of TODOs is the issue
tracker?

Closes: #1222
Approved by: cgwalters