scripts: Replace Fedora rawhide glibc's lua %transfiletriggerin
See: https://pagure.io/atomic-wg/issue/419

In Fedora rawhide glibc changed to use `<lua>` for running `ldconfig`.
I experimented with different changes to the glibc package; for
example, using `-p /sbin/ldconfig` but the problem is that passes
arguments which `ldconfig` errors out on; we'd have to do something
like add `/usr/libexec/ldconfig-for-rpm` as a hardlink and have it
dispatch behavior on `argv[0]`.

The problem with simply using `/bin/sh` is it creates a circular
dependency today, even though really by the time we're doing `%posttrans`
we *always* have a shell, so the dependency is silly.  Down the
line we should investigate teaching RPM to omit `/bin/sh` dependencies
at least for `%posttrans`.