WIP: status: Display RpmMdRepos key
I think the set of enabled rpm-md repos is of fundamental
importance.  Let's display it in status.

There's quite a lot of subtle things going on here...one that
is IMO confusing is: this PR displays the *current* repos which
may be different from the rpm-mdrepos used for the base commit (included
by default), which may also be different from the repos used to generate
the *current* deployment.

Another way to say this is most of the status entries are about the deployment,
and are basically immutable, but this is not. I'm not sure how to handle that,
but it does feel weird.

Perhaps instead we show the repos that were actually used to generate
the layered set, and highlight any diffs vs current?