scripts: Disable rofiles-fuse for glibc locales
OK so I give up on this glibc locales issue; see
https://github.com/ostreedev/ostree/pull/1470

Based on what I've been reading it looks like newer FUSE may
help, but we can't rely on that.

We should also rewrite the upstream glibc code to be less pathological; it'd
make sense to use `O_TMPFILE` for example, and keep track of how many bytes we
wrote rather than calling `fstat()`, and rather than `truncate()` on the tmpl
file at the end, do a create-empty-file-then-`rename()`. But for now let's go
with this so we get the pretty `●` back in rojig mode!

Closes: #1280
Approved by: jlebon