app: fix wrong EXIT_FAILURE returns
Fix all instances where we were returning `EXIT_FAILURE` instead of
`FALSE`. Also be more precise and use `gboolean` instead of `int` as
return type, which probably makes it easier to fall in this trap.

We should probably add a `make syntax-check` or something to scan for
these.

Closes: #1294
Approved by: cgwalters