bwrap: Use user namespaces and avoid overmounting
Now that we've split out the "base" constructor for the initramfs
special case, the "mutable" disposition tells us how to handle
binding `/etc` and `/var` in addition to `/usr`.  Drop the
`RPMOSTREE_BWRAP_IMMUTABLE` case as it is now only used in the
script case where we handle `/var`.

Next, user namespaces denies overmounting existing targets as it
could reveal things intended to be hidden.  In the recursive
container case, we need to just use e.g. `--bind /dev /dev`
rather than trying to create a new instances.

Further, in the selftest case we can't overmount what already
exists, so skip doing that entirely.

Closes: https://github.com/projectatomic/rpm-ostree/issues/1329