core,scripts: When no cachedir is enabled, disable FUSE use
This is also prep for running inside (unprivileged) Kube containers
as they exist today: https://github.com/projectatomic/rpm-ostree/issues/1329

Sadly FUSE today uses a suid binary that ends up wanting CAP_SYS_ADMIN.
I think there's some work on FUSE-in-containers but I'm not sure of
the current status.

What FUSE is protecting is the hardlinked repo imports.  But if
`--cachedir` isn't specified (which we support today) that repository
gets thrown away anyways.  So there's no real value to using FUSE
here.  Let's hence make it so that in this scenario we don't.

Down the line ideally we gain the capability to detect if either
unprivileged overlayfs/FUSE are available.  Then if `--cachedir`
is specified we can make things work.