status: Show rpmmd-repo data in verbose mode
Building on:

 - 9cbec27d4c69b27afdaf895f84b64a4f0a3de3d6
 - e7a42f70a9e119f5211665f8ed036e8f82fd34e2

I was looking at a rpm-ostree run that imports a variety of rpmmd-repos,
and information about the source repositories is really useful for determining
the up-to-dateness.  We've been capturing this data for a while, it's
about time we started showing it somewhere.

This does make `status --verbose` notably more verbose, but eh, that
seems fine for now.

See also https://github.com/projectatomic/rpm-ostree/issues/774