syscore: Only do a single repo prune for base+pkgcache
Since we unified the pkgcache repo in 7056e6b726a804446c28a03f6752eab662abb010
there's no longer a reason to perform two repository prunes.

Change things so that the first phase is regenerating all of the
refs (in a single libostree txn), then perform a single prune.

This is preparation for reworking how we do prunes, which is
going to be useful for staged deployments.